detail content 수정 버튼 disable 및 로딩  적용
