+++ src/components/common/Loading/styles/backgroundLoading.module.css
@@ -11,7 +11,7 @@
 .loadingSection {
   z-index: 999;
   position: fixed;
-  top: 54%;
+  top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
   justify-content: center;
