+++ src/pages/schedule/DetailRoom/TitleSection.tsx
@@ -3,12 +3,12 @@ import styles from './style/detailSchedule.module.css';
 import { IDateDetail, IMemberInfo } from '../../../types/room';
 import Input from '../../../components/common/Input/Input';
 import Button from '../../../components/common/Button/Button';
-import { updateDetailDateContentsApi } from '../../../services/room/room';
 import Toast from '../../../components/common/Toast/Toast';
 import Span from '../../../components/common/Span/Span';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { detailScheduleInfo, roomInfo } from '../../../recoil/room/roomInfo';
-import { aa } from './dododo';
+import BackgroundLoading from '../../../components/common/Loading/BackgroundLoading';
+import { updateContentsFunc } from '../../../utils/room/updateContents';
 
 interface IProps {
   myProfile: IMemberInfo;
@@ -20,6 +20,7 @@ const TitleSection = ({ myProfile }: IProps) => {
   const room = useRecoilValue(roomInfo);
 
   const [visible, setVisible] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
   const [toastText, setToastText] = useState('');
 
   const [newTitle, setNewTitle] = useState(detailSchedule.subTitle);
@@ -36,16 +37,17 @@ const TitleSection = ({ myProfile }: IProps) => {
   };
 
   const updateTitleHandler = async () => {
-    // newTitle을 갖고 있는 detailDate
+    setIsLoading(true);
     const newDetail: IDateDetail = {
       ...detailSchedule,
       subTitle: newTitle,
     };
 
-    aa(room, detailSchedule.id, newDetail, () => {
+    updateContentsFunc(room, detailSchedule.id, newDetail, () => {
       setDetailSchedule({ ...detailSchedule, subTitle: newTitle });
       subTitleToggle();
       setVisible(!visible);
+      setIsLoading(false);
       setToastText('수정되었습니다!');
     });
   };
@@ -73,6 +75,7 @@ const TitleSection = ({ myProfile }: IProps) => {
             onClick={updateTitleHandler}
             text={'수정'}
             className={'newTitleUpdateBtn'}
+            disable={isLoading}
           />
           <Button
             onClick={subTitleToggle}
@@ -86,6 +89,7 @@ const TitleSection = ({ myProfile }: IProps) => {
       {visible && (
         <Toast text={toastText} visible={visible} setVisible={setVisible} />
       )}
+      {isLoading && <BackgroundLoading />}
     </div>
   );
 };
